fix: [spg-849]: Imposing locks on audit cleanup job. (#39832)

+++ 400-rest/src/main/java/software/wings/service/impl/instance/AuditCleanupJob.java
@@ -7,6 +7,9 @@
 
 package software.wings.service.impl.instance;
 
+import io.harness.lock.AcquiredLock;
+import io.harness.lock.PersistentLocker;
+
 import software.wings.service.intfc.AuditService;
 
 import com.google.common.annotations.VisibleForTesting;
@@ -14,6 +17,7 @@ import com.google.common.util.concurrent.ThreadFactoryBuilder;
 import com.google.inject.Inject;
 import com.google.inject.Singleton;
 import io.dropwizard.lifecycle.Managed;
+import java.time.Duration;
 import java.time.Instant;
 import java.time.LocalDateTime;
 import java.time.ZoneOffset;
@@ -29,7 +33,9 @@ public class AuditCleanupJob implements Managed {
   private static final long DELAY_IN_MINUTES = TimeUnit.HOURS.toMinutes(24);
 
   private static int retentionTimeInMonths = 18;
+  private static final String LOCK_NAME = "AUDIT_CLEANUP_JOB_LOCK";
   @Inject private AuditService auditService;
+  @Inject private PersistentLocker persistentLocker;
 
   private ScheduledExecutorService executorService;
 
@@ -50,9 +56,16 @@ public class AuditCleanupJob implements Managed {
 
   @VisibleForTesting
   public void run() {
-    log.info("Audit Cleanup Job Started @ {}", Instant.now());
-    long toBeDeletedTillTimestamp =
-        LocalDateTime.now().minusMonths(retentionTimeInMonths).toInstant(ZoneOffset.UTC).toEpochMilli();
-    auditService.deleteAuditRecords(toBeDeletedTillTimestamp);
+    try (AcquiredLock<?> lock =
+             persistentLocker.tryToAcquireInfiniteLockWithPeriodicRefresh(LOCK_NAME, Duration.ofSeconds(10))) {
+      if (lock == null) {
+        log.info(LOCK_NAME + "failed to acquire lock");
+        return;
+      }
+      log.info("Audit Cleanup Job Started @ {}", Instant.now());
+      long toBeDeletedTillTimestamp =
+          LocalDateTime.now().minusMonths(retentionTimeInMonths).toInstant(ZoneOffset.UTC).toEpochMilli();
+      auditService.deleteAuditRecords(toBeDeletedTillTimestamp);
+    }
   }
 }
